pyproject.toml
[tool.poetry]
name = "couchbase-boostrap"
version = "0.1.0"
description = ""
authors = ["Sven Ulland <sven.ulland@gmail.com>"]
readme = "README.md"
packages = [{include = "couchbase_bootstrap"}]

[tool.poetry.dependencies]
python = "^3.8"
python-json-logger = "^2.0.4"
PyYAML = "^6.0"
couchbase-cluster-admin = { git = "https://github.com/sveniu/couchbase-cluster-admin.git" }
backoff = "^2.2.1"

[tool.poetry.scripts]
couchbase-bootstrap = "couchbase_bootstrap.cli:cli"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.122"
black = "^22.10.0"
jmespath = "^1.0.1"
pytest = "^7.2.0"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
